refactor: recipe 용어 summary 로 변경
