+++ src/store/selectors/summarySelector/cartSummarySelector.test.ts
@@ -6,7 +6,7 @@ import { RecoilRoot, useRecoilValue, useSetRecoilState } from "recoil";
 import { cartListState, filteredCartItemState } from "@/store/atoms/atoms";
 
 import MOCK_CART_LIST from "@/constants/_mock/mockCartList";
-import { recipeState } from "@/store/selectors/recipeSelector/recipeSelector";
+import { cartSummaryState } from "@/store/selectors/summarySelector/cartSummarySelector";
 import { renderHook } from "@testing-library/react";
 
 jest.mock("../../../api/config", () => ({
@@ -30,7 +30,7 @@ describe("cartTotalPriceState", () => {
         });
 
         const { orderPrice, shippingFee, totalPrice } =
-          useRecoilValue(recipeState);
+          useRecoilValue(cartSummaryState);
 
         return { orderPrice, shippingFee, totalPrice };
       },
@@ -58,7 +58,7 @@ describe("cartTotalPriceState", () => {
         });
 
         const { orderPrice, shippingFee, totalPrice } =
-          useRecoilValue(recipeState);
+          useRecoilValue(cartSummaryState);
 
         return { orderPrice, shippingFee, totalPrice };
       },
